Flask API - generates Python3 bindings for headers
Use the new FFIG build_from_source way of passing unsaved_files. Receives a json with a header file and returns json with binding. 

Tested with a requests module that sends a test input and uses difflib to confirm the generated binding is as expected

Minor: tidy-up Dockerfile, update python and fix pythonpath